Cycle count feature, three ways to execute and reconfiguration templates
----------------------------------------------------------------------
Reconfiguration templates are the following:
0x3210 -> 4 contexts with 2 pipelanes each.
0x0211-> 3 contexts with 4 lanes in Ctx 1, 2 lanes Ctx 2, 2 lanes ctx 0.
0x0011-> 2 contexts, with 4 cores for each context (Ctx 1 and zero).
0x0001 or 0x0000 -> Default configuration (1 context with 8 pipelanes).

The behaviour is basically as it follows:
Each 'Y' on the number '0xYYYY' is intended to represent a context and has 2 pipelanes in it. So if you want to turn on 2 cores with 4 pipelanes in each, you'd do '0x0011' or '0x1122', and so on. These rules apply if you want to turn on only 3 cores  ,etc. 
----------------------------------------------------------------------
The cycle count feature is embbedded in the 'testbenchwithcyclecount.vhd' file, so it can run a different command from the interface. Check the file if you want more informations about the code (it basically counts each cycle for each lane and sum all context to show the total cycles from one application).
----------------------------------------------------------------------
The three ways to execute are:
1) make vsim-APPLICATION
2) make vsimrun-APPLICATION
3) make vsimruncyclecount-APPLICATION
1 is the basic simulation with Modelsim interface.
   [You need to execute the 'do runwave.do' command in Modelsim interface to simulate and see the waves.]
2 is the basic simulation without Modelsim interface (only console).
3 is the simulation without Modelsim interface and with Cycle counting.