Execution in parallel
Parallel execution is measured with the amount of time both programs (executing in 4-issue at the same time) spent. The code for executing in parallel can be found in the 'parallel.c' file at /test-progs/src/parallel.c

The logic behind executing them in parallel can be found in the code itself. 
Gains of 9-11% are made when the ucbqsort and jpeg benchmarks are executed in parallel than executing them on a separated environment (4-issue and 8-issue alone).

The measurement was made with the 'make vsimruncyclecount-parallel' command.